+++ apps/coop_chat/src/main.rs
@@ -11,11 +11,11 @@ mod generated_code {
 
 pub use generated_code::*;
 
-pub mod controller;
-pub mod mocks;
-mod view_controller;
+// pub mod controller;
+// pub mod mocks;
+// mod view_controller;
 
-use view_controller::*;
+// use view_controller::*;
 
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen(start))]
 pub fn main() {
@@ -24,23 +24,13 @@ pub fn main() {
     #[cfg(all(debug_assertions, target_arch = "wasm32"))]
     console_error_panic_hook::set_once();
 
-    #[cfg(feature = "slint_orbclient")]
-    slint_orbclient::init_config(
-        slint_orbclient::Config::default()
-            .width(1000)
-            .height(600)
-            .resizable(true)
-            .events_async(true)
-            .title("coop_chat"),
-    );
-
     #[cfg(feature = "slint_coop")]
     slint_coop::init_config(600., 400., "coop_chat");
 
     let app = App::new();
 
-    let _side_bar_view_controller =
-        SideBarViewController::new(&app, Box::new(mocks::MockSideBarController::new()));
+    // let _side_bar_view_controller =
+    //     SideBarViewController::new(&app, Box::new(mocks::MockSideBarController::new()));
 
     app.run();
 }