+++ apps/coop_chat/src/view_controller.rs
@@ -1,6 +1,6 @@
-// SPDX-FileCopyrightText: 2022 Florian Blasius <co_sl@tutanota.com>
-// SPDX-License-Identifier: GPL-3.0-only
+// // SPDX-FileCopyrightText: 2022 Florian Blasius <co_sl@tutanota.com>
+// // SPDX-License-Identifier: GPL-3.0-only
 
-mod side_bar_view_controller;
+// mod side_bar_view_controller;
 
-pub use side_bar_view_controller::*;
+// pub use side_bar_view_controller::*;