+++ gaming/pico_designer/src/main.rs
@@ -1,6 +1,0 @@
-// SPDX-FileCopyrightText: 2022 Florian Blasius <co_sl@tutanota.com>
-// SPDX-License-Identifier: GPL-3.0-only
-
-fn main() {
-    println!("Hello, world!");
-}