+++ widgets/book_flip/src/docs.rs
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2022 Florian Blasius <co_sl@tutanota.com>
+// SPDX-License-Identifier: MIT
+
+#![cfg(doc)]
+
+/*!
+    This is only a helper to generate the documentation for the Slint components.
+*/
+
+pub mod widgets {
+    #![doc = include_str!("../ui/_widgets.md")]
+}